requirements-sidecar.txt
# Perception Sidecar — Python dependencies
# These run ONLY in the sidecar container, never in Open WebUI.

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0
pyyaml>=6.0
Pillow>=10.0
httpx>=0.25.0

# Florence-2 inference
torch>=2.1.0
transformers>=4.36.0
einops
timm

# HuggingFace model downloads
huggingface-hub>=0.19.0

# Optional: WD-14 tagger (only needed if wd14 backend is enabled)
# pandas
# onnxruntime-gpu  # alternative to torch for some taggers
